fix(ci): repair the Go verification gates and clear what they were hiding (#345)
ENG-021. Three related defects let the Go gates report a clean tree that was
not clean.

Loop exit status. Both the pre-commit golangci-lint hook and the Makefile's
deadcode branch looped over go.work modules in a shell loop whose status is
the status of its LAST iteration, so a failing module was masked by any later
module that passed. Not hypothetical: apps/collector had 9 findings and
apps/notifier 3, and the hook still exited 0 because apps/market-hotset, the
last module, is clean.

Parser drift. The hook parsed go.work with grep '^use ', which yields an
empty list against the valid use (...) block form and then lints nothing and
succeeds. That is the drift go_workspace_modules.sh was written to end; this
hook was simply never migrated to it.

Config schema. .golangci.yml declared version '2' while writing
linters-settings at the root and exclude-dirs under issues, both v1
locations. golangci-lint run ignores unknown keys silently, so every setting
read as configured and none of it applied.

Both loops now live in scripts beside that parser: module list from it, fail
on an empty list, run every module, fail if any failed. go_lint.sh verifies
the config schema before linting anything. The hook also triggers on the
config and the wrappers themselves, so an invalid key cannot reach main in a
commit touching no Go file.

gosec severity/confidence moves to its v2 location, where it changes nothing.
gocyclo.min-complexity 15 and gocritic's diagnostic/performance/style tags
are deliberately not restored: dead for as long as this file has been on v2,
and applying them surfaces 161 findings including eight production functions
needing real refactoring. That policy decision is ENG-029 with the measured
counts, not a side effect of repairing the gate.

Twelve findings the masked hook hid are fixed: a dropped rdb.Close error,
three integer overflow conversions on counters that only count upward (now
uint64 throughout), an if-else chain, a suspicious compound condition
rewritten with the same truth table, a preallocated slice, De Morgan on a
negated conjunction, a test helper parameter that only ever received "bybit",
and a test whose twenty sequential assertions crossed gocyclo's default,
now one table with the same coverage. Two nilerr reports are intentional and
now say why: a cancelled context is a normal shutdown, and one malformed
liquidation payload is counted and dropped rather than tearing down the
capture stream.

Tests run the real scripts against a fake golangci-lint/deadcode and a
throwaway repo root: a failure injected into the first, middle or last module
fails the gate while every module is still visited; both go.work forms give
the same module list; an empty workspace and an invalid config fail before
anything is linted. Six of them fail against the previous one-liner.

+++ apps/collector/internal/liquidationcapture/writer.go
@@ -95,9 +95,13 @@ func (w *Writer) Flush(ctx context.Context) error {
 			batch.Queue(insertEventSQL, eventArgs(event)...)
 		}
 		results := w.db.SendBatch(flushCtx, batch)
-		insertedCount := 0
-		duplicateCount := 0
-		mismatchCount := 0
+		// uint64 rather than int plus a conversion at the accumulation
+		// below: these only ever count upward from zero into uint64 stats,
+		// so the conversion was both unnecessary and an integer-overflow
+		// finding (gosec G115).
+		insertedCount := uint64(0)
+		duplicateCount := uint64(0)
+		mismatchCount := uint64(0)
 		for _, event := range batchEvents {
 			var inserted bool
 			var storedHash []byte
@@ -129,9 +133,9 @@ func (w *Writer) Flush(ctx context.Context) error {
 
 		w.mu.Lock()
 		w.pending = w.pending[n:]
-		w.stats.EventsPersistedTotal += uint64(insertedCount)
-		w.stats.DuplicateEventsTotal += uint64(duplicateCount)
-		w.stats.PayloadHashMismatchTotal += uint64(mismatchCount)
+		w.stats.EventsPersistedTotal += insertedCount
+		w.stats.DuplicateEventsTotal += duplicateCount
+		w.stats.PayloadHashMismatchTotal += mismatchCount
 		w.stats.LastPersistAt = time.Now()
 		w.mu.Unlock()
 		remaining -= n
